Refactor to use tokens
Slight change - instead of storing copies of strings everywhere now I
just use "tokens," ie indices to a dictionary of words from the original
text. The code is still ugly as sin but PROGRESS

Also added a to-do list so I can keep track of what stuff I need to work
on.